[debugger] Parse nested Rust tuple accesses.
Rust supports the syntax "tuple.0" (or a nonzero number) to access the
corresponding value of the Tuple. Zxdb supported this as a special case
and not a float because in Rust, numbers can't start with a dot.

But Zxdb's parsing still allowed the following digits to be parsed as a
float. So if there were nested tuple accesses like "tuple.0.1" it was
parsed as <"tuple"> <dot> <float=0.1>.

This adds a new context-sensitive mode to the tokenizer for Rust where
seeing a dot without a leading digit actively inhibits treating the
following characters as a float.

Change-Id: I45504801f3c3de091c050078c4c71b401d0121ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003815
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>