[netstack3] Support sending to broadcast addresses
Supports sending Ipv4 broadcast traffic to the Ethernet
all-ones broadcast address.

This includes both the limited broadcast IP address (255.255.255.255)
and the subnet broadcast address (e.g. for 192.168.0.0/24,
192.168.0.255).

As a result, this causes the out-of-stack DHCP client to be able to
successfully Rebind on netstack3 (as it depended on sending UDP
broadcast traffic in order to do so).

Fixed: 42148833
Change-Id: Ib61620049147c718b9b02b0f3a54e9922da44ca4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970713
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>