[netstack3] Update the bug # referenced by a TODO
This is the last TODO referring to TCP dual stack sockets. The nature
of this TODO is not critical to the functionality and should be a
cleanup. Update the reference so that we can mark the umbrella bug as
closed.

Fixed: 42085913
Change-Id: I15bb1d48c2ed115a316d053c76d7ab61d9c07e44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003259
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>