[phys] Ensure physboot-elf-kernel-test has uncompressed kernel package
As an interim measure, physload reuses the space previously occupied by
the kernel package (i.e., the STORAGE_KERNEL ZBI item) for the hand-off
to the kernel. With profile instrumentation, this hand-off further
encodes profiles which can take up a fair amount of space. When the
kernel package is compressed, it is too small to fit a hand-off with
profiles, so we reserve more space by leaving it uncompressed for now.

We also address a cosmetic image metadata bug that would incorrectly
present uncompressed ZBIs as compressed.

Bug: 325513615
Change-Id: I0d987758f419fd6d58cd14468b5813c496da294c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993241
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>