[amlogic_decoder] Fix bind tests and build rules
The build rules for the amlogic_decoder driver bind rules were organized
such that the tests were only defined in the host toolchain and only
when the target_cpu (within the host toolchain) was set to arm64. In
practice this meant the tests only ran on arm64 hosts (which are quite
rare) and as a result the bind test have bitrotted. This moves the build
rules for the bind tests out of the conditional, since they should pass
on any host architecture, and updates the test expectations to match the
current logic. The test now passes on arm64 and x64 hosts.

Bug: 324437929
Bug: 283475308
Change-Id: I36abd58e03278d3a15eff1945d6773597d4b3718
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/999673
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>