[starnix/usercopy] Ret slice backed by dest slice
When `Usercopy::copyin{,_until_null_byte}` sees an address that is not
in the restricted address range, it terminates early and returns an
empty slice that it creates. The problem here is that callers of
the copyin methods assume that the returned initialized slice has the
same address as the uninitialized slice provided to the method.

This change fixes the copyin methods to always return an initialized
slice that has the same address as the provided uninitialized slice.

Bug: b/328610022
Change-Id: I1573dd3c3e8538a5559d88bf570c7829ba5c238f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004134
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>