[amlogic-display][viu] Do not pass canvas index as an argument of FlipOnVsync.
Previously the canvas_idx field of the display image was always
passed in to FlipOnVsync() to flip the active frame buffer used
by the VIU.

However, the canvas_idx field is actually invalid (and unused) when
AFBC is enabled. A better way to get the field is by directly
checking the `info` struct, and only when the image is not AFBC-
compressed.

This change fixes issue by removing the canvas index argument
from the FlipOnVsync() method.

Test: no-op.
Test: on VIM3 with AML028-30MB-A1, "display-tool squares" works.
      Note that "display-tool squares" generates non-AFBC frame
      buffers, so the modified code is tested.
Bug: 323060335
Change-Id: I7b93d508f8ac32556fec9b79ad97db182accff11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004234
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>