[pkg-cache] Change how the dynamic index will be disabled
The dynamic index is still enabled, this just changes how the flag to
disable it works.

Previously, the dynamic index was disabled by ignoring it during GC.
This does not work, because the dynamic index verifies state
transitions during Gets that will start to fail when we use the open
package cache instead of the dynamic index to short-circuit Gets.

Now we disable the dynamic index by, when the flag is set, having all
mutating operations on it do nothing, so it will always be empty.

Bug: 327208811
Change-Id: I721fa14538e5ef0815dfe7b9414e06bee2c39fa8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004133
Reviewed-by: Sen Jiang <senj@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>