[bedrock] Introduce libavailability
It is intended to be a minimal library used by libsandbox
when Router is moved there.

librouting also depends on it to serve legacy routing.

While doing this I discovered that OfferFromVoidToRequiredTarget is in
accurate, and is not used by bedrock. I didn't remove this out of
concern that legacy routing might be relying on this error to stop
routing early.

Bug: 324870669
Change-Id: I6147c3e0307c7f0cd3eced561b2fd8f177ff4e7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004135
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>