[test][power] Fix a regression in the power capture scripts.
A small regression crept in last Friday when change
Ie23e6d081e3c54a2b2d0d25267d29881a9a2eef6 landed.  Specifically, this:

```
result = proc.wait(60)
if result:
  <handle error>
```

became this:

```
result = proc.wait(60)
if result is not None:
  <handle error>
```

`proc`, in this case, is a Popen object from the subprocess library.
We have just signaled to the subprocess we spawned that it should
terminate.  We then wait for up to 60 seconds for the process to
terminate, and then check our error code to see if something went
wrong.

`Popen.wait(timeout)` is always going to do one of two things.  It
will either throw an exception, because the process did not return
within the timeout, or it will return the return code of the process
which just exited.  It will never return `None` (`poll()` can do that,
but not `wait()`).  By convention, a return value of 0 means that the
process had no errors, and a non-zero return value means that
something went wrong.

The previous code was testing an integer against zero, and if non-zero
was handling an error (a C-style predicate).  The new form, testing
against `None`, probably seemed to be more explicit, but unfortunately
it was incorrect in this situation.

Fix the issue by changing the code to explicitly test for
non-zero-ness instead of testing for none-None-ness.

Change-Id: I2deab691ccfdf6f820cf536d2a6821f41f8a4a40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005648
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Commit-Queue: John Grossman <johngro@google.com>