[netstack2] Add logging for Terminal(User)
Add an overabundance of logging to help track down the cause of the
"Terminal(User)" flakes.  The goal is to log the key flow control paths
of interface creation and teardown to understand the order of operations
that's leading to the failure.  I expect this is far more than is
necessary, but I don't want to have to go through more rounds of this if
I can avoid it.

The logging should also be inexpensive enough to not cause issues.
These RPCs are administrative, so there aren't tight performance
bounds, and we're only adding a handful of syscalls.  We're also not
adding any new failure points.

Bug: 328297563
Change-Id: I1cf1e7f7e3a2f7ae23d725d41273808cd4f28a1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005852
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>