[fdf][devfs] Separate Devnode Connectors
As part of removing fidl multiplexing, separate
the Connector callbacks for connecting to the
fuchsia.device/Controller channel and the device
channel. Devnode now takes both callbacks, and
calls the appropriate one depending on what file
the devfs client opens.  This eliminates the possibility
of any new implementations accidentally multiplexing
fidl channels, and clarifies to connection process.

Note that the compat shim still uses the device channel
to connect to the controller, so a hack is added to route
the compat shims connections correctly.  This is eliminated
in a follow-up CL.

Bug: 42063787

Testing: tests are updated to test the new devnode interface.

Change-Id: I005282133527aeeb1729846f484c0ed36463ce6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005016
Reviewed-by: Nick Eaton <nickeaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Garratt Gallagher <garratt@google.com>