[bt][obex] Assign a ConnectionId if requested
The remote OBEX client may request a directed OBEX session to a specific
Service by including the Target Header with UUID in the CONNECT request.

Assign a Connection ID to the connection if requested. This ID will be
included in all subsequent OBEX operations and is fixed for the lifetime
of the OBEX connection. Because there is only one OBEX server for an
underlying RFCOMM/L2CAP transport, a static ID of 1 is assigned. This is
mainly a formality and is irrelevant in almost all situations.

Bug: b/42080293
Test: fx test bt-obex-tests
Change-Id: Ida8e962fc486739dfead4a0fa6635e510dd9f996
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005372
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>