[netstack3] Fix socket inspect local addr any
Print the address correctly for TCP sockets that are locally bound to
the ANY address.

As `SocketAddr`'s type definition forces an address to be specified,
it was necessary to generalize to allow the unspecified ANY address.
Added `AddrAndPortFormatter` for formatting an address and a port, and
reuse this implementation in `SocketAddr`. Modified `AddrAndZone` and
`ZonedAddr` where necessary in support of this.

Modified the existing socket integration test to assert this.

Fixed: 326283980
Change-Id: I3b071c432dd46769767a42c5f25c1f4d2e3ec316
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1002858
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>