[display][banjo] New struct for ImportImage() args.
DisplayController.ImportImage() currently reuses the Image structure to
convey the expected usage of the buffers in a sysmem BufferCollection.
This is misleading, because one of the fields (the imported image handle) is
not meaningful.

This CL reduces the confusion by adding a dedicated struct for
ImportImage(), and by updating the method comment to clarify our
expectation that this information will be sourced from sysmem at some
point in the future.

Bug: 316631158
Change-Id: I19a2a7c47998a64503dde1217bb12c51925f52a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1004841
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
API-Review: John Bauman <jbauman@google.com>