[logs] Add ordered mutex
For both the old and new symbolizer we want to
ensure that futures are polled in a particular
order at a synchronization point, and data
at that point is guarded. This is similar
to a futures::lock::Mutex but is different
in that it is optimized for single-threaded
contexts and guarantees tasks are unblocked in the
order that they called lock.

MULTIPLY: host_x64/ffx_daemon_log_lib_test

Bug: b/299980894
Change-Id: I84eb48f6d6db0d816784ad9ac9a8617e6e79bf65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006876
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>