[fxtest][debug] Wait for debugger to be ready
zxdb has a small startup time to connect to the backend on the target
and install any filters for test components that will be spawned.
There's a race condition between TestManager spawning all of the test
cases and DebugAgent such that a test that fails sufficiently fast may
not be caught by the debugger, or worse, crash the debugger due to some
strange state that would otherwise be impossible.

This makes use of the existing support for zxdb's signal-when-ready
argument.

Bug: b/325281869
Change-Id: I494975c281fa5a7a6f04fea4b98be59cd24845b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1006057
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>