[zircon] zx_sampler_*
This CL Proposes/Creates a @next, build argument gated, syscall enabling
sampling data from user threads.

Currently, a user program can sample a given thread by obtaining a
handle, suspending the thread, then using process read memory to iterate
over the targets stack. This process takes ~2-4ms on supported hardware
and is too slow to be useful.

This CL introduces an approach for the kernel to provide assistance for
taking samples. Creating a sampler causes the kernel to allocate and
return a buffer to the user that it will write samples into. The caller
can then request a thread start being sampled. This causes the kernel to
(when safe), sample the thread's user stack and write it to the buffer.
With this kernel assistance, sampling overhead is now ~2-4us per sample.

These syscalls are intended as a temporary solution to unblock
performant profiling on Fuchsia as well as a building block and
iteration opportunity to design and build a robust set of observability
apis for Fuchsia.

Bug: b/301632554
Change-Id: Ieabf0e3031a82bc08a799e34ca3fc4f67d079522
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939358
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
API-Review: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Meggin Kearney <mkearney@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>