[zircon] Add zx_sampler_* syscalls
Adds a group of @next only syscall to mark a thread to have its stack
periodically sampled and written to a buffer.

The general flow looks like:
- Use zx_sampler_create to set up a session and get buffers that will be
  written to
- Call zx_sampler_attach on each relevant thread to enable sampling
- Take some samples
- Call zx_sampler_stop to disable sampling
- Map the assigned buffers and read out the written data

Calling zx_sampler_attach  marks a thread to be sampled
periodically as described in a passed in config. This comes with some
complications since sampling the thread requires reading user memory
which isn't always safe to do from a timer context.

Rather than immediately attempting to read the thread's user stack, we
set a new thread signal `THREAD_SIGNAL_SAMPLE_STACK`. This indicates to
the thread that the next time it's in a safe spot to do so, it should
walk its stack and write it out to the buffer assigned to its core.

Bug: b/301632554
Change-Id: I09c5d2115cb0ded5b1bcb80ed9fee079b40a4780
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933215
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>