[ffx][sdk] Adding machine writer to `sdk version`
This adds the use of VerifiedMachineWriter to the sdk version
subcommand.

This is part of a project to make all commands that are integrated
into other programs/scripts use the machine interface and provide
a schema of the output to enable identifying breaking changes.

Bug: b/324937400
Change-Id: Ib506495ceb2968539a9ee465062bf98bb5ab13ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1005637
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>