[sdk][ffx] Add the next round of separate subtools
This adds separate binaries for the next set of subtools that should be
moved out of the main ffx binary into the SDK. This needs to go in
before it's removed from ffx to make sure there's some overlap.

The actual removal CLs will be done separately to allow for easy
reverts.

Bug: b/329326263
Change-Id: Id518971e086e339191a47bf5084e7b4fe1592bbb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959798
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>