[graphics] Use fdio_service_connect to connect to Provider
We need to stop sending OnOpen events for service connections as part of
the io2 migration, as this is no longer supported in io2. This is
because it requires clients switch protocols over the service channel by
first listening for a fuchsia.io/OnOpen event, then switching protocols.

The VFS libraries now ensure an epitaph is sent on failed service
connections, which clients can listen for or handle as if the service
was closed.

Bug: b/327261966
Test: fx test get_hardware_display_coordinator_with_provider_service_test get_hardware_display_coordinator_without_provider_service_test
Change-Id: I53e603d1361821fcf720cb216a0e6581cb85268e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1003772
Reviewed-by: Yilong Li <liyl@google.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>