[rfc] Asynchronous operations are on objects
This proposes changing the semantics of asynchronous operations in the
kernel to be defined as operating on objects as opposed to the handles
referring to those objects. This modifies the behavior of the
`object_wait_async` syscall when the handle is closed and proposes a
new syscall for cancelling all asynchronous operations registered with a
given key.

Change-Id: Ic32a5848c1026feaf8637cfdca5808c1f88ce7bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985309
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Dave Moore <davemoore@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>