[forensics] Generate a unified binary & rename it
This patch reorganizes the forensics build so that it generates a unified binary, and uses renamed_binary template to generate a list of renamed_binary targets. The problem was originally discovered in the referred bug, but resurfaced in the LTO context, where we started
seeing a size increase because of duplication in forensics binaries.

Bug: 323231755
Change-Id: I097da0c73e2d646797d0577a87a841794db5c180
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007796
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>