[zircon] Fix providing ZX_HANDLE_INVALID to process_start
...when the bad handle policy is enabled. This system call is defined to
allow providing the value ZX_HANDLE_INVALID if no handle transfer is
required but the handle lookups used to implement this would trigger the
bad handle policy if enabled. This guards the handle manipulations so
that in the ZX_HANDLE_INVALID case no policy errors are raised.

This codepath is tested by ProcessTest.ProcessStartNoHandle in the core
tests and the Zircon Rust library wrapper tests but we don't currently
run either test suite with the bad handle policy enabled so these tests
do not report an error. We should run these suites (or subsets of them)
with the bad handle policy set to a higher enforcement level.

Change-Id: I0726d6c8850b667a585a455f90e069f0c7281be5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007714
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>