[network/bench] Wait for interface control handle closure
When cleaning up after running the "interfaces" workload in the resource
usage benchmarks, wait for the control handle to the interface to be
closed with a terminal event rather than just dropping the control
handle and waiting to observe removal on the interfaces/State protocol.
Netstack3 does not wait for all interface cleanup work to be complete
before emitting the interface removal event to State watchers, but it
does not emit terminal events on the control handle until all cleanup is
complete.

Synchronizing only on the State removal event meant that the benchmark
observed that resource usage had not subsided to baseline for Netstack3
even after the benchmark had finished running.

Fixed: 322860526
Test: resource-usage-benchmark-netstack3
Change-Id: I3f8c1509c667acdbbd372ede79cd81f1d27e9723
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1007294
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>